RWA-1143 fixed failing tests due to DMN changes
